refactored ReaderViewPagerTransformer into WPViewPagerTransformer and moved into widgets folder
